Move image loading to view model
